Выполнил ДЗ по модульности
